+++ include/highfive/bits/H5Annotate_traits_misc.hpp
@@ -51,7 +51,7 @@ AnnotateTraits<Derivate>::createAttribute(const std::string& attribute_name,
 template <typename Derivate>
 template <typename Type, typename InferredType>
 inline Attribute
-AnnotateTraits<Derivate>::writeAttribute(const std::string& attribute_name,
+AnnotateTraits<Derivate>::createAttribute(const std::string& attribute_name,
                                          const InferredType& data) {
     Attribute att = createAttribute(attribute_name, DataSpace::From(data), AtomicType<Type>());
     att.write(data);